Refactor building visit history timeline
