+++ server/constants/requestMethods.ts
@@ -12,12 +12,11 @@ export const requestMethodsBooking: Readonly<Partial<Record<ApplicationMethodTyp
   BY_PRISONER: 'By the prisoner',
 }
 
-export const requestMethodDescriptions: Readonly<Record<ApplicationMethodType, string>> = {
+// Used in visit details timeline
+export const requestMethodDescriptions: Partial<Record<ApplicationMethodType, string>> = {
   PHONE: 'Method: Phone booking',
   WEBSITE: 'Method: GOV.UK request',
   EMAIL: 'Method: Email request',
   IN_PERSON: 'Method: In person booking',
   BY_PRISONER: 'Method: Prisoner request',
-  NOT_KNOWN: '', // Return empty string when request method is unknown
-  NOT_APPLICABLE: '', // Return empty string when request method is not applicable
 }
