+++ server/services/visitService.ts
@@ -4,7 +4,6 @@ import {
   ApplicationDto,
   ApplicationMethodType,
   CancelVisitOrchestrationDto,
-  EventAudit,
   Visit,
   VisitBookingDetailsDto,
   VisitPreview,
@@ -12,21 +11,6 @@ import {
 import { HmppsAuthClient, OrchestrationApiClient, RestClientBuilder } from '../data'
 import logger from '../../logger'
 import { prisonerDateTimePretty, prisonerTimePretty } from '../utils/utils'
-import eventAuditTypes from '../constants/eventAudit'
-import { requestMethodDescriptions } from '../constants/requestMethods'
-import { notificationTypes } from '../constants/notifications'
-
-export type MojTimelineItem = {
-  label: { text: string }
-  text: string
-  datetime: {
-    timestamp: string
-    type: 'datetime'
-  }
-  // byline undefined is necessary when no 'user' is present for "by User X"
-  byline: { text: string }
-  attributes: { 'data-test': string }
-}
 
 export default class VisitService {
   constructor(
@@ -215,80 +199,6 @@ export default class VisitService {
     const orchestrationApiClient = this.orchestrationApiClientFactory(token)
 
     return orchestrationApiClient.getBookedVisitCountByDate(prisonId, date)
-  }
-
-  getVisitEventsTimeline({
-    events,
-    visitStatus,
-    visitNotes,
-  }: {
-    events: EventAudit[]
-    visitStatus: VisitBookingDetailsDto['visitStatus']
-    visitNotes: VisitBookingDetailsDto['visitNotes']
-  }): MojTimelineItem[] {
-    const filteredEvents = events.filter(event => Object.keys(eventAuditTypes).includes(event.type)).reverse()
-
-    let cancelledVisitReason = ''
-    if (visitStatus === 'CANCELLED') {
-      visitNotes.forEach(note => {
-        if (note.type === 'VISIT_OUTCOMES') {
-          cancelledVisitReason = note.text
-        }
-      })
-    }
-
-    return filteredEvents.map((event, index) => {
-      const label = eventAuditTypes[event.type]
-
-      let descriptionContent = ''
-      const { applicationMethodType, userType, type } = event
-
-      if (type === 'BOOKED_VISIT' || type === 'UPDATED_VISIT' || type === 'MIGRATED_VISIT') {
-        if (userType === 'PUBLIC' && applicationMethodType === 'WEBSITE') {
-          descriptionContent = 'Method: GOV.UK booking'
-        } else {
-          descriptionContent = requestMethodDescriptions[event.applicationMethodType]
-        }
-      } else if (type === 'CANCELLED_VISIT') {
-        if (cancelledVisitReason !== '') {
-          descriptionContent = `Reason: ${cancelledVisitReason}`
-        } else if (applicationMethodType !== 'NOT_KNOWN' && applicationMethodType !== 'NOT_APPLICABLE') {
-          if (userType === 'PUBLIC' && applicationMethodType === 'WEBSITE') {
-            descriptionContent = 'Method: GOV.UK cancellation'
-          } else {
-            descriptionContent = requestMethodDescriptions[applicationMethodType]
-          }
-        }
-      } else if (type === 'IGNORE_VISIT_NOTIFICATIONS_EVENT') {
-        descriptionContent = `Reason: ${event.text}`
-      } else if (
-        type === 'PRISONER_RELEASED_EVENT' ||
-        type === 'PRISON_VISITS_BLOCKED_FOR_DATE' ||
-        type === 'PRISONER_RECEIVED_EVENT'
-      ) {
-        // only added to assist type for next line
-        descriptionContent = `Reason: ${notificationTypes[type]}`
-      }
-
-      const user = event.actionedByFullName
-
-      let byline = null
-      if (user && user !== 'NOT_KNOWN') {
-        if (user === 'NOT_KNOWN_NOMIS') {
-          byline = { text: 'NOMIS' }
-        } else {
-          byline = { text: user }
-        }
-      }
-
-      return {
-        label: { text: label }, //
-        text: descriptionContent,
-        datetime: { timestamp: event.createTimestamp, type: 'datetime' }, //
-        byline, //
-        attributes: { 'data-test': `timeline-entry-${index}` },
-      }
-    })
   }
 
   private buildVisitInformation(visit: Visit): VisitInformation {
